# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AlphaAI Backend API",
    "version": "1.0.0",
    "description": "API documentation for the AlphaAI Backend service",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key needed to access the endpoints"
      }
    },
    "schemas": {
      "Person": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "title": {
            "type": "string"
          },
          "companyId": {
            "type": "integer"
          },
          "researchStatus": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "fullName",
          "email",
          "companyId"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "The specified resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/people": {
      "get": {
        "summary": "Get all people",
        "tags": [
          "People"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of people",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "summary": "Create a new person",
        "tags": [
          "People"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "full_name",
                  "email",
                  "company_id"
                ],
                "properties": {
                  "full_name": {
                    "type": "string",
                    "description": "Full name of the person"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address"
                  },
                  "company_id": {
                    "type": "integer",
                    "description": "ID of the company"
                  },
                  "title": {
                    "type": "string",
                    "description": "Job title"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Person created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Company not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/enrich/{person_id}": {
      "post": {
        "summary": "Trigger enrichment research for a person",
        "tags": [
          "People"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID of the person to enrich"
          }
        ],
        "responses": {
          "200": {
            "description": "Research started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "jobId": {
                      "type": "string"
                    },
                    "person": {
                      "$ref": "#/components/schemas/Person"
                    },
                    "company": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Person must be associated with a company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Person not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/research/jobs/{job_id}": {
      "get": {
        "summary": "Get the status of a research job",
        "tags": [
          "Research"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the job to check"
          }
        ],
        "responses": {
          "200": {
            "description": "Job status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string"
                    },
                    "progress": {
                      "type": "number"
                    },
                    "data": {
                      "type": "object"
                    },
                    "result": {
                      "type": "object"
                    },
                    "timestamps": {
                      "type": "object",
                      "properties": {
                        "created": {
                          "type": "number"
                        },
                        "started": {
                          "type": "number"
                        },
                        "finished": {
                          "type": "number"
                        }
                      }
                    },
                    "attempts": {
                      "type": "object",
                      "properties": {
                        "current": {
                          "type": "number"
                        },
                        "max": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/research/status": {
      "put": {
        "summary": "Bulk update research status for multiple people",
        "tags": [
          "Research"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "personIds": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    }
                  },
                  "status": {
                    "type": "string"
                  }
                },
                "required": [
                  "personIds",
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Research status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
...
}